% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clip_trials.R
\name{clip_trials}
\alias{clip_trials}
\title{Clip Trials to Standard Length}
\usage{
clip_trials(data, trial, time, start, stop)
}
\arguments{
\item{data}{Data frame containing all of your time series data and metadata.}

\item{trial}{Name of the variable containing the trial label.}

\item{time}{Name of the variable containing the time stamp.}

\item{start}{Value for start time. Times earlier than this will be discarded.}

\item{stop}{Value for stop time. Times later than this will be discarded.}
}
\value{
Copy of data frame reduced to the range of times given by
\code{start} and \code{stop}.
}
\description{
Clips all trials to the specified start and stop times, and issues a warning
if there are any missing samples.
}

