% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_artifacts.R
\name{fix_artifacts}
\alias{fix_artifacts}
\title{Artifact correction}
\usage{
fix_artifacts(ts, samp_freq, artifacts = NULL, baseline = NULL,
  lim = c(-0.5, 0.5), max_gap = 1, ...)
}
\arguments{
\item{ts}{A time series, passed as a vector of chronologically ordered
observations separated by equal intervals of time.}

\item{samp_freq}{The sampling frequency in Hz.}

\item{artifacts}{A logical vector of equal length to the time series that
provides logical indexing into which entries of the time series correspond
to artifacts. If this vector is not specified, it will be generated using
the \code{\link{get_artifacts}} function.}

\item{baseline}{Logical vector indicating which parts of the time series
correspond to the baseline period. This is used as a reference for
thresholding the interpolations to the limits of relative change set in the
\code{lim} argument. If there is no baseline, the mean of the entire series
(excluding signal loss) will be used for this reference.}

\item{lim}{Vector of limits specifying minimum and maximum relative change
from baseline that is plausible for your experiment. Default is
\code{c(-.5,.5)}, meaning inteprolations will be constrained to fall within
a 50\% decrease or 50\% increase from \code{baseline}.}

\item{max_gap}{The maximum missing time interval in seconds to fill. Any
longer gaps will be left unchanged. The default setting is 1 sec.}

\item{...}{Additional arguments to be passed to \code{\link{get_artifacts}}}
}
\value{
A copy of the time series with artifacts and missing data replaced
by interpolated values.
}
\description{
\code{fix_artifacts} Corrects artifacts using a sequence of linear
interpolation for internal artifacts (artifacts sandwiched between good data)
followed by matched lag-1 differences to overwrite external artifacts
(when the trial starts or ends with an artifact).
}
\section{Warning}{

If the time series contains gaps that are too long for interpolation, the
function will silently return a time series consisting of all missing values
so that the user can easily identify which trials need to be dropped.
}
\seealso{
\code{\link{get_artifacts}}, \code{\link[zoo]{na.approx}}
}

