% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_artifacts.R
\name{get_artifacts}
\alias{get_artifacts}
\title{Artifact detection}
\usage{
get_artifacts(ts, samp_freq, baseline = NULL, lim = c(-0.5, 0.5),
  max_loss = 0.5, max_gap = 1, min_cont = 0.1, max_velocity = NULL)
}
\arguments{
\item{ts}{A time series, passed as a vector of chronologically ordered
observations separated by equal intervals of time.}

\item{samp_freq}{Sampling frequency in Hz.}

\item{baseline}{Logical vector indicating which parts of the time series
correspond to the baseline period. This is used as a reference for
determining whether values exceed the limits of relative change set in the
\code{lim} argument. If there is no baseline, the mean of the entire series
(excluding signal loss) will be used for this reference.}

\item{lim}{Vector of limits specifying minimum and maximum relative change
from baseline that is plausible for your experiment. Default is
\code{c(-.5,.5)}, meaning values that represent more than a 50\% decrease or
50\% increase from baseline will be considered artifacts.}

\item{max_loss}{Maximum fraction of total signal loss that is tolerable.
Default value is 0.5, meaning time series with more than 50\% signal loss will
be rejected.}

\item{max_gap}{Maximum gap in seconds that a loss of signal is allowed to
span. Default value is 1 second.}

\item{min_cont}{Minimum continuity in seconds required between artifacts.
If the period of "good" data between two artifacts is less than this
threshold, the two artifacts are merged into one. Default value is 0.1
(100 ms).}

\item{max_velocity}{The maximum plausible change (in absolute diameter units)
that can occur from one reading to the next. Differences between consecutive
readings that exceed this value are used to identify artifact onsets, and
differences within this value over the \code{min_cont} time span are used to
identify artifact offsets. If no value is specified, the function will assign
this to the 90th percentile of the absolute values of the first-order lag-1
differences pooled across the entire time series.}
}
\value{
A logical vector that can be used for logical indexing into the time
series to identify data artifacts.
}
\description{
\code{get_artifacts} returns a logical vector indicating the location of all
artifacts present in a pupil-diameter time series.
}
\details{
This algorithm is an extension of other blink-detection algorithms and seeks
to identify all artifacts--not just blinks--that can appear when continuously
measuring pupil diameter over time.

Like other algorithms, this one relies primarily on abrupt jumps in signal to
identify the onset and offset of blinks. Unlike other algorithms, it does not
require the user to prespecify threshold values that define onsets and
offsets for all time series; rather, it adaptively determines the best values
based on the distribution of the first-order differences (velocities) of each
time series. This procedure was designed to mimic the relativistic way a
human observer would visually identify an artifact, i.e., by assessing the
pattern of deviation for the candidate artifact relative to that of its
neighbors. However, this function can also accommodate the traditional
approach by assigning static thresholds via optional parameters.
}
\section{Disclaimer}{

This algorithm is intended for the detection of artifacts in relatively short
time series corresponding to pupil-dilation measurements. It has only been
tested and validated on 8-second trials sampled at 60 Hz or 500 Hz. Good
artifact detection may or may not generalize to other sampling rates, trial
lengths, and types of data.
}

