% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_freq.R
\name{get_freq}
\alias{get_freq}
\title{Calculate Sampling Frequency}
\usage{
get_freq(time, units = "ms")
}
\arguments{
\item{time}{Numeric vector of time stamps.}

\item{units}{String indicating units of time (either \code{"ms"} or
\code{"s"}). The default is \code{"ms"}.}
}
\value{
Sampling frequency in Hz.
}
\description{
If you are unsure about the sampling frequency of your time series,
\code{get_freq} can verify this, provided your data contains a time stamp for
each measurement.
}
