% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tlbs.R
\name{get_tlbs}
\alias{get_tlbs}
\title{Trial level bias score}
\usage{
get_tlbs(measure, type, reference = NULL, search_limit = 5)
}
\arguments{
\item{measure}{A vector of chronologically ordered observations of numeric
type.}

\item{type}{A vector of equal length to \code{measure} that provides the
trial type of the corresponding entry of \code{measure}. Must be a
dichotomous factor (having exactly 2 levels) or convertible to such.}

\item{reference}{Value indicating which level of trial \code{type} should
come first when subtracting trials of opposite type to generate a bias score
(bias = \code{reference} type - opposite type). If \code{type} is of class
\code{factor}, defaults to the first level; otherwise defaults to the trial
type that comes first alphanumerically.}

\item{search_limit}{How many trials to look forward or backward to find a
trial of opposite type. Default value is 5. If no match is found within the
\code{search_limit} of a trial, \code{NA} will be returned for that trial.}
}
\value{
A vector of bias scores for each trial.
}
\description{
\code{get_tlbs} Given an ordered series of trials of dichotomous type,
each trial is matched with the trial of opposite type that is most temporally
proximal, and the difference score is returned.
}

