% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_outlier.R
\name{is_outlier}
\alias{is_outlier}
\title{Identify Outliers}
\usage{
is_outlier(measure, abs_lim = NULL, sd_lim = NULL)
}
\arguments{
\item{measure}{A numeric vector.}

\item{abs_lim}{Two-item numeric vector \code{c(lower, upper)}
specifying the \emph{absolute} lower and upper limits, respectively, of the
range of valid values in terms of \emph{measurement units}. This range should
be set such that values that fall outside of it are considered implausible or
impossible.}

\item{sd_lim}{Numeric value specifying the \emph{relative} range of valid
values in terms of \emph{standard deviation}.}
}
\value{
A logical vector indicating \code{TRUE} if the trial is outside the
given range of valid values.
}
\description{
\code{is_outlier} flags which observations fall outside a valid range of
values based on limits set by the user. Limits can be set in absolute
terms (the units of measurement), standard deviations, or both.
}
\details{
If both absolute and standard-deviation limits are specified, then
\code{is_outlier} will first apply the absolute-deviation test, followed
by the standard-deviation test for only those observations that pass the
absolute-deviation test (so that values that are outright impossible do not
factor into the determination of standard deviation).

\code{is_outlier} uses a robust version of median absolute deviation to
estimate the population standard deviation, which remains valid even for
non-normal and/or asymmetric distributions. See \href{
http://eurekastatistics.com/using-the-median-absolute-deviation-to-find-outliers}{
Peter Rosenmai's blog post} for more information.
}
\examples{
# Create example time series of 10 reaction times in ms with two trials that
# fall outside the bounds of validity:
set.seed(7)
rt <- c(0, 10000, rnorm(8, mean = 1000, sd = 250))
rt
# Check for trials that are less than 250 ms with no upper bound:
is_outlier(rt, abs_lim = c(250, Inf))

# Check for trials that are more than 2 standard deviations from the mean
is_outlier(rt, sd_lim = 2)

# Check for trials that are less than 250 ms, more than 2500 ms, or more than
# 2 standard deviations from the mean:
is_outlier(rt, abs_lim = c(250, 2500), sd_lim = 2)

}
\references{
Rosemani, P. 2013. \href{
http://eurekastatistics.com/using-the-median-absolute-deviation-to-find-outliers}{
Using the median absolute deviation to find outliers.}
}

