% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_valid.R
\name{is_valid}
\alias{is_valid}
\title{Valid trials.}
\usage{
is_valid(measure, limit, type)
}
\arguments{
\item{measure}{A vector of chronologically ordered observations of numeric
type.}

\item{limit}{A two-item numeric vector \code{c(lower, upper)} specifying the
lower and upper limits, respectively, of the range of valid values. The
range can be specified in terms of either "raw" or "sd" (standard deviation)
values, or a mixture of the two as defined by the \code{type} argument.}

\item{type}{A one- or two-item vector corresponding to the type of unit
("raw" or "sd") of the \code{limit} argument. If the lower and upper
\code{limit} are specified in different units, then this argument should
correspond to \code{c(type_lower, type_upper)}.}
}
\value{
A logical vector indicating \code{TRUE} if the trial is valid and
 \code{FALSE} if it is not.
}
\description{
\code{is_valid} indicates which trials fall within a valid range of values.
}
\examples{
# Create example time series of 10 reaction times in ms with two trials that
# fall outside the bounds of validity:
rt <- c(0, 20000, sample(1000:5000, 8))

# Check that trials are not less than 250 ms and not more than 2.5 standard
# deviations above the mean:
is_valid(rt, limit = c(250, 2.5), type = c("raw","sd"))

}

