% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_eyes.R
\name{merge_eyes}
\alias{merge_eyes}
\title{Merge Eyes}
\usage{
merge_eyes(left, right)
}
\arguments{
\item{left}{Numeric vector containing the time series of measurements for the
left eye.}

\item{right}{Numeric vector containing the time series of measurements for
the right eye.}
}
\value{
Numeric vector. For binocular readings, the mean time series for
both eyes. For monocular readings, the time series for the eye that was
tracked.
}
\description{
Aggregate parallel time series from the left and right eyes into a single
time series of measurements.
}

