% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\title{Normalize Time Series}
\usage{
normalize(ts, baseline)
}
\arguments{
\item{ts}{A time series, passed as a vector of chronologically ordered
observations separated by equal intervals of time.}

\item{baseline}{A logical vector indicating which elements of the time
series correspond to the baseline period.}
}
\value{
A baseline-normalized copy of the time series converted to units of
relative change (difference from baseline as a fraction of baseline).
}
\description{
\code{normalize} normalizes the time series to a specified baseline
reference period using \eqn{(ts-baseline)/baseline} where \eqn{ts} is the
time series and \eqn{baseline} is the mean of the values belonging to the
baseline period. If the baseline period is too unstable (usually an
indication that \code{\link{get_artifacts}} failed to detect an artifact
during the baseline period and that its tuning parameters should be
adjusted), a warning will be issued and normalization will not be performed.
(A vector of \code{NAs} will be returned.)
}
\section{Warning}{

Interpolation of missing values should be done prior to running this
function. Time series with any \code{NA}s will be silently rejected by
returning a vector consisting entirely of \code{NA}s.
}

\seealso{
\code{\link{artifacts}}, \code{\link{low_pass_filter}}
}
