% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out_of_range.R
\name{out_of_range}
\alias{out_of_range}
\title{Identify Out-of-Range Observations}
\usage{
out_of_range(measure, abs_limit = NULL, sd_limit = NULL)
}
\arguments{
\item{measure}{A vector of chronologically ordered observations of numeric
type.}

\item{abs_limit}{Optional two-item numeric vector \code{c(lower, upper)}
specifying the \emph{absolute} lower and upper limits, respectively, of the
range of valid values in terms of \emph{measurement units}. This range should
be set such that values that fall outside of it are considered implausible or
impossible.}

\item{sd_limit}{Numeric value specifying the \emph{relative} range of valid
values in terms of \emph{standard deviation}. This should be set such that
values that fall outside of +/- \code{sd_limit} from the mean are considered
outliers. Defaults to 3 SD if not specified.}
}
\value{
A logical vector indicating \code{TRUE} if the trial is outside the
given range of valid values.
}
\description{
\code{out_of_range} flags which observations fall outside a valid range of
values based on limits set by the user. Limits can be set in absolute
terms (the units of measurement), standard deviations from the mean, or both.
If both types of limit are specified, then \code{out_of_range} will
first apply the absolute-deviation test, followed by the standard-deviation
test for only those observations that pass the absolute-deviation test (so
that values that are outright impossible do not factor into the calculation
of the mean and standard deviation).
}
\examples{
# Create example time series of 10 reaction times in ms with two trials that
# fall outside the bounds of validity:
set.seed(7)
rt <- c(0, 10000, rnorm(8, mean = 1000, sd = 250))
rt
# Check for trials that are less than 250 ms with no upper bound:
out_of_range(rt, abs_limit = c(250, Inf))

# Check for trials that are more than 2 standard deviations from the mean
out_of_range(rt, sd_limit = 2)

# Check for trials that are less than 250 ms, more than 2500 ms, or more than
# 2 standard deviations from the mean:
out_of_range(rt, abs_limit = c(250, 2500), sd_limit = 2)

}

