% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_comparison.R
\name{plot_comparison}
\alias{plot_comparison}
\title{Plot Comparison of Original vs. Cleaned Time Series}
\usage{
plot_comparison(data, time, orig, clean, trial = NULL)
}
\arguments{
\item{data}{Data frame containing both time series.}

\item{time}{Name of variable that gives time units.}

\item{orig}{Name of variable containing the original (uncleaned) time series
of measures.}

\item{clean}{Name of variable containing the cleaned time series of
measures.}

\item{trial}{If the data contains multiple time series, name of the variable
that identifies each time series. Will be plotted using \code{facet_wrap}.}
}
\description{
\code{plot_comparison} helps to visualize the transformation of your original
time series into a subsequent, cleaner version.
}
\details{
Use this plotting function to evaluate the performance of the
\code{\link{fix_artifacts}}, \code{\link{normalize}}, and/or
\code{\link{low_pass_filter}} functions on cleaning your time series. Note
that if your orignal and cleaned time series are at different scales (e.g.,
if you are comparing a pre-normed version to a post-normed version),
the original time series will be projected to the scale of
the cleaned time series.
}

