% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relabel_trial.R
\name{relabel_trial}
\alias{relabel_trial}
\title{Relabel trial numbers}
\usage{
relabel_trial(trial, valid)
}
\description{
Assign invalid trials and renumber remaining trials as a consecutive sequence
from 1 to the total number of trials.
}
