% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artifacts.R
\name{artifacts}
\alias{artifacts}
\alias{fix_artifacts}
\alias{get_artifacts}
\title{Time Series Artifact Detection and Correction}
\usage{
get_artifacts(ts, samp_freq, lim, baseline = NULL, max_loss = 0.5,
  max_gap = 1, min_cont = 0.1, max_velocity = NULL)

fix_artifacts(ts, samp_freq, lim, baseline = NULL, max_loss = 0.5,
  max_gap = 1, ..., threshold = TRUE, artifacts = NULL)
}
\arguments{
\item{ts}{A time series, passed as a numeric vector of chronologically
ordered, positively valued observations separated by equal intervals of time.}

\item{samp_freq}{Sampling frequency in Hz.}

\item{lim}{Two-item numeric vector \code{c(neg, pos)} specifying the negative
and positive limits, respectively, of relative change from baseline
that is plausible for your time series. (See section "Advice on setting
\code{lim} argument" for more information and examples.) Values outside of
this range will be flagged as artifacts regardless of whether or not they
deviate markedly from their nearest neighbors.}

\item{baseline}{Logical vector indicating which parts of the time series
correspond to the baseline period. This is used as a reference for
determining whether values exceed the limits of relative change set in the
\code{lim} argument. If there is no \code{baseline} provided, the mean of the
entire series (excluding signal loss) will be used for this reference.}

\item{max_loss}{Maximum fraction of time series that is allowed to contain
dropped signal and/or artifacts. Default value is 0.5, meaning if more than
half of the time series is missing, \code{get_artifacts} will label the
entire time series as one continuous artifact, and, if more than half of the
time series consists of artifacts, \code{fix_artifacts} will not perform
interpolation/extrapolation and instead will return a vector of all
\code{NA}s.}

\item{max_gap}{Maximum gap in seconds that an artifact period is allowed to
span. Default value is 1 second.}

\item{min_cont}{Minimum continuity in seconds required between artifacts.
If the period of "good" data between two artifacts is less than this
threshold, the two artifacts are merged into one. Default value is 0.1
(100 ms).}

\item{max_velocity}{The maximum plausible change (in measurement units)
that can occur from one reading to the next. Differences between consecutive
readings that exceed this value are used to identify artifact onsets, and
differences within this value over the \code{min_cont} time span are used to
identify artifact offsets. If no value is specified, the function will assign
this to the 90th percentile of the absolute values of the first-order lag-x
differences (where x is the number of samples that best approximates 20 ms
given the \code{samp_freq}) pooled across the entire time series.}

\item{...}{Additional arguments to be passed to \code{get_artifacts} via
\code{fix_artifacts}}

\item{threshold}{Logical value indicating whether or not thresholding to the
range specified in \code{lim} should be performed when extrapolating the
beginning and ending of a time series (if the time series begins or ends with
an artifact). If set to \code{FALSE}, extrapolations that exceed \code{lim}
will cause the entire time series to be rejected and a vector of all
\code{NA}s to be returned. If set to \code{TRUE}, the lower and upper
\code{lim} will be used to impose a floor and ceiling, respectively, on
extrapolation.}

\item{artifacts}{A logical vector of equal length to the time series that
provides logical indexing into which entries of the time series correspond
to artifacts. This can be obtained by first calling \code{get_artifacts}. If
not supplied by the user, then \code{fix_artifacts} will call
\code{get_artifacts} itself.}
}
\value{
\code{get_artifacts} returns a logical vector that can be used for
logical indexing into the time series to identify data artifacts.
\code{fix_artifacts} returns a copy of the time series with artifacts and
missing data replaced by interpolated values, or a copy of the time series
with all values changed to \code{NA} in the event that the artifacts are too
numerous (exceed \code{max_loss}) or too continuous (exceed \code{max_gap}).
}
\description{
\code{get_artifacts} and \code{fix_artifacts} identify and repair,
respectively, signal artifacts within a time series. They are designed to
work with pupil measurements obtained from an eye-tracking system, primarily
to identify and interpolate over eye blinks, but also any other type of
signal artifact that can appear when continuously measuring pupil diameter or
area over time.
}
\details{
Like similar algorithms, \code{get_artifacts} relies primarily on abrupt
jumps in signal to identify the onset and offset of blinks. Unlike other
algorithms, it does not require the user to prespecify threshold values that
define onsets and offsets for all time series; rather, it adaptively
determines the best values based on the distribution of the first-order
differences (velocities) of each time series. This procedure was designed to
mimic the relativistic way a human observer would visually identify an
artifact, i.e., by assessing the pattern of deviation for the candidate
artifact relative to that of its nearest neighbors. However, this function
can also accommodate the traditional approach by assigning static thresholds
via optional parameters.

Using the output of \code{get_artifacts}, \code{fix_artifacts} repairs
artifacts using a sequence of linear interpolation for internal artifacts
(artifacts sandwiched between good data) followed by matched lag-1
differences for external artifacts (artifacts at the start or end of a time
series).
}
\section{Advice on setting \code{lim} argument}{

You are required to specify a vector of limits that defines what magnitude of
relative change from baseline is plausible for your experiment and your
equipment. \code{get_artifacts} uses the \code{lim} argument to catch periods
of sustained signal loss or signal drift that would not be caught by looking
at velocity alone, e.g., if the signal slowly (rather than abruptly) veers
into a range that is implausible given its starting point, and
\code{fix_artifacts} uses the \code{lim} argument to constrain extrapolations
at the beginning and end of trials. Depending on the \code{threshold}
argument, \code{fix_artifacts} imposes a floor or ceiling to the
extrapolation (when \code{threshold = TRUE}), or rejects the entire time
series (when \code{threshold = FALSE}).

In setting the \code{lim} argument, think about what sort of relative change
is plausible for your measurement. For example, if you are measuring pupils,
the normal pupil size in adults varies from a minimum diameter of about 2 mm
(3 square mm area) to a maximum diameter of about 8 mm (50 square mm area) in
bright light vs. darkness. This means if the pupil starts off at maximum
dilation, it can experience at most a 75\% decrease in diameter (95\%
reduction in area). If the pupil starts off at minimum dilation, it can
experience at most a 300\% increase in diameter (1500\% increase in area).
This would correspond to \code{lim = c(-0.75, 3)} for diameter and
\code{lim = c(-0.95, 1500)} for area, but you should set your limits to be
more conservative if you do not expect your measurements to span this range.

For example, most users will measure pupils under moderate lighting
conditions, so baseline pupil readings will start off closer to the center of
their physiological range. Even assuming maximum decreases and increases from
this point, the range could be narrowed to \code{lim = c(-0.6, 0.6)} for
diameter and \code{lim = c(-0.85, 1.5)} for area. We have found that with
psychological stimuli using our eye-tracking setup, \code{lim = c(-0.5, 0.5)}
and \code{lim = c(-.75, 1.25)} appear to provide liberal coverage for
plausible changes in pupil diameter and area, respectively, and you want your
\code{lim} setting to err on the side of being too wide. The \code{lim}
argument is meant to serve as an "extra layer of security" to catch a
minority of artifacts that slip past the primary detection algorithm. So a
narrower `lim` will generate far more false positives than a wider `lim` will
generate false negatives.
}

\section{Warning}{

If the time series contains gaps that are too long for interpolation (as
defined in the \code{max_gap} argument) or if the extrapolated portion of
the time series exceeds the range specified in the \code{lim} argument and
\code{threshold} is set to \code{FALSE}, \code{fix_artifacts}
will silently return a time series consisting of all missing values so that
the user can easily identify which trials need to be dropped.
}

\section{Disclaimer}{

This algorithm is intended for the detection of artifacts in relatively short
time series corresponding to pupil-dilation measurements. It has only been
tested and validated on 6-8 second trials sampled at 60 Hz or 500 Hz. Good
artifact detection may or may not generalize to other sampling rates, trial
lengths, and types of data. Please use \code{\link{plot_artifacts}} to
inspect the performance of \code{get_artifacts} before continuing with
\code{fix_artifacts} and any subsequent data cleaning and analysis.
}
\seealso{
\code{\link{plot_artifacts}},  \code{\link{plot_comparison}},
 \code{\link{normalize}},  \code{\link{low_pass_filter}}
}

