% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bs.R
\name{get_bs}
\alias{get_bs}
\title{Mean bias score}
\usage{
get_bs(RT, congruent)
}
\arguments{
\item{RT}{A numeric vector of reaction times in chronological order.}

\item{congruent}{A logical vector equal in length to \code{RT} that indicates
whether the corresponding entry of \code{RT} is a congruent \code{TRUE} or
incongruent \code{FALSE} trial.}
}
\value{
A length-one numeric vector.
}
\description{
\code{get_bs} returns the mean difference between trials of dichotomous
type.
}
\examples{
# Create example time series of 10 reaction times in ms:
rt <- sample(100:1000, 10)

# Create example trial types of congruent vs. incongruent for above measures:
congruent <- sample(c(TRUE,FALSE), 10, replace = TRUE)

get_bs(rt, congruent)

}
\seealso{
\code{\link{get_tlbs}} for trial-level bias score
}

