% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bs.R
\name{get_bs}
\alias{get_bs}
\title{Mean bias score}
\usage{
get_bs(RT, congruent, prior_weights = NULL)
}
\arguments{
\item{RT}{A numeric vector of reaction times in chronological order.}

\item{congruent}{A logical vector equal in length to \code{RT} that indicates
whether the corresponding entry of \code{RT} is a congruent (\code{TRUE}) or
incongruent (\code{FALSE}) trial, i.e., whether the probe location matches
the location of the stimulus expected to elicit attention bias.}

\item{prior_weights}{Optional numeric vector of prior weights indicating the
relative influence that each trial should have on the calculation of
mean bias, in the event that there is reason to weight some trials more than
others when calculating their mean.}
}
\value{
A length-one numeric vector corresponding to the mean, or optionally
 weighted mean, bias score.
}
\description{
\code{get_bs} returns the mean difference between congruent and incongruent
trials.
}
\details{
Attention bias tasks such as the dot probe consist of congruent trials
(CTs), in which the location of the probe matches the location of an
emotional stimulus, and incongruent trials (ITs), in which the location of
the probe matches the location of the neutral stimulus. \code{get_bs}
calculates a traditional bias score by taking the mean reaction time of all
CTs and subtracting it from the mean reaction time of all ITs, i.e.,
\eqn{bias = IT - CT}, with the option to use weighted means based on prior
trial weights.
}
\examples{
# Create example time series of 10 reaction times in ms:
rt <- sample(100:1000, 10)

# Create example trial types of congruent vs. incongruent for above measures:
congruent <- sample(c(TRUE,FALSE), 10, replace = TRUE)

get_bs(rt, congruent)

}
\seealso{
\code{\link{get_tlbs}} for trial-level bias score
}

