% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_outlier.R
\name{is_outlier}
\alias{is_outlier}
\title{Identify Outliers}
\usage{
is_outlier(measure, abs_lim = NULL, mad_lim = NULL, sd_lim = NULL)
}
\arguments{
\item{measure}{A numeric vector.}

\item{abs_lim}{Two-item numeric vector \code{c(lower, upper)}
specifying the \emph{absolute} lower and upper limits, respectively, of the
range of valid values in terms of \emph{measurement units}. This range should
be set such that values that fall outside of it are considered implausible or
impossible.}

\item{mad_lim}{Numeric value specifying the range of valid values in terms of
\emph{median absolute deviations} from the median.}

\item{sd_lim}{Numeric value specifying the range of valid values in terms of
\emph{standard deviations} from the mean.}
}
\value{
A logical vector indicating \code{TRUE} if an observation is an
 outlier.
}
\description{
\code{is_outlier} flags which observations fall outside a valid range of
values based on limits set by the user. Limits can be set in absolute
terms (the units of measurement), median absolute deviations, standard
deviations, or any combination of the above.
}
\details{
If more than one type of limit is specified, \code{is_outlier} will first
apply the absolute limits if given (so that values that are outright
impossible do not factor into the determination of the deviation statistics),
followed by the median-absolute-deviation (MAD) test and/or the
standard-deviation test.

The \code{mad_lim} argument is evaluated using the \emph{double MAD}, which
provides for robust identification of outliers even when the underlying
distribution is non-normal and/or asymmetric. See \href{
http://eurekastatistics.com/using-the-median-absolute-deviation-to-find-outliers}{
Peter Rosenmai's blog post} for more information.
}
\examples{
# Create example time series of 10 reaction times in ms with two trials that
# fall outside the bounds of validity:
set.seed(7)
rt <- c(0, 10000, rnorm(8, mean = 1000, sd = 250))
rt
# Check for trials that are less than 250 ms with no upper bound:
is_outlier(rt, abs_lim = c(250, Inf))

# Check for trials that are more than 3 standard deviations from the mean
is_outlier(rt, sd_lim = 3)

# Check for trials that are more than 3 median absolute deviations from the
# median
is_outlier(rt, mad_lim = 3)

# Check for trials that are less than 250 ms, more than 2500 ms, or more than
# 2.5 MADs from the median:
is_outlier(rt, abs_lim = c(250, 2500), mad_lim = 2.5)

}
\references{
Rosenmai P. 2013. \href{
http://eurekastatistics.com/using-the-median-absolute-deviation-to-find-outliers}{
Using the median absolute deviation to find outliers.}
}

