% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_artifacts.R
\name{plot_artifacts}
\alias{plot_artifacts}
\title{Plot artifacts}
\usage{
plot_artifacts(time, pupil, artifacts, data = NULL, ...)
}
\arguments{
\item{time}{Numeric vector corresponding to time. Will be plotted on the x-
axis. Must be the same length as the \code{pupil} and \code{artifact}
vectors.}

\item{pupil}{Numeric vector containing the time series of pupil measurements.
Will be plotted on the y-axis. Must be the same length as the \code{time}
and \code{artifact} vectors.}

\item{artifacts}{Logical vector that indexes which observations in the
\code{pupil} time series are artifacts. See \code{\link{get_artifacts}}.}

\item{data}{Data frame to use (optional).}

\item{...}{optional arguments to \code{\link[ggplot2]{qplot}},
e.g., "main", "ylim", etc.}
}
\description{
\code{plot_artifacts} returns a plot of a single time series with
artifact intervals marked in red with vertical dotted lines.
}
\details{
Use this plotting function to evaluate the performance of the
\code{\link{get_artifacts}} function on individual time series and the
impact of changes to its threshold settings. If many artifacts are
going undetected, try decreasing the \code{max_velocity} setting. If there
are too many false positives, try increaasing it.
}

