% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_comparison.R
\name{plot_comparison}
\alias{plot_comparison}
\title{Plot Comparison of Time Series}
\usage{
plot_comparison(data, time, pre, post, trial = NULL)
}
\arguments{
\item{data}{Data frame containing both time series.}

\item{time}{Name of variable that gives time units.}

\item{pre}{Name of variable containing the pre-processed time series.}

\item{post}{Name of variable containing the post-processed time series.}

\item{trial}{If the data contains multiple time series, name of the variable
that identifies each time series. Will be plotted using \code{facet_wrap}.}
}
\description{
\code{plot_comparison} helps to visualize the transformation of your original
time series into a subsequent, cleaner version.
}
\details{
Use this plotting function to evaluate the performance of the
\code{\link{fix_artifacts}}, \code{\link{normalize}}, and/or
\code{\link{low_pass_filter}} functions on cleaning your time series. Note
that if your original and cleaned time series are at different scales (e.g.,
if you are comparing a pre-normed version to a post-normed version),
the original time series will be projected to the scale of
the cleaned time series.
}
