% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_onset.R
\name{zero_onset}
\alias{zero_onset}
\title{Zero Time Index to Trial Onset}
\usage{
zero_onset(time, marker, onset)
}
\arguments{
\item{time}{Numeric vector of sample indicies or time stamps corresponding to
a single time series.}

\item{marker}{Vector containing the indicator of trial \code{onset}.}

\item{onset}{Value of \code{marker} that indicates when the trial begins.}
}
\value{
Numeric vector that zeros \code{time} to the trial onset.
}
\description{
Centers the sample index or time stamp of your time series to a reference
marker, such that 0 always coincides with the onset of a trial.
}

